address code review comments (#2)
