+++ src/gui/tsconfig.composite.json
@@ -5,6 +5,5 @@
 	],
 	"include": [
 		"./**/*.ts",
-		"../api/**/*.ts"
 	]
 }
