+++ src/model/chart-model.ts
@@ -1,9 +1,5 @@
 /// <reference types="_build-time-constants" />
 
-import {
-	MouseEventHandlerEventBase,
-} from '../gui/mouse-event-handler';
-
 import { assert, ensureNotNull } from '../helpers/assertions';
 import { gradientColorAtPercent } from '../helpers/color';
 import { Delegate } from '../helpers/delegate';
@@ -29,6 +25,7 @@ import { Series, SeriesOptionsInternal } from './series';
 import { SeriesOptionsMap, SeriesType } from './series-options';
 import { LogicalRange, TimePointIndex, TimeScalePoint } from './time-data';
 import { TimeScale, TimeScaleOptions } from './time-scale';
+import { TouchMouseEventData } from './touch-mouse-event-data';
 import { Watermark, WatermarkOptions } from './watermark';
 
 /**
@@ -354,7 +351,7 @@ export class ChartModel implements IDestroyable {
 	private _width: number = 0;
 	private _hoveredSource: HoveredSource | null = null;
 	private readonly _priceScalesOptionsChanged: Delegate = new Delegate();
-	private _crosshairMoved: Delegate<TimePointIndex | null, MouseEventHandlerEventBase | null> = new Delegate();
+	private _crosshairMoved: Delegate<TimePointIndex | null, TouchMouseEventData | null> = new Delegate();
 
 	private _backgroundTopColor: string;
 	private _backgroundBottomColor: string;
@@ -493,7 +490,7 @@ export class ChartModel implements IDestroyable {
 		return this._crosshair;
 	}
 
-	public crosshairMoved(): ISubscription<TimePointIndex | null, MouseEventHandlerEventBase | null> {
+	public crosshairMoved(): ISubscription<TimePointIndex | null, TouchMouseEventData | null> {
 		return this._crosshairMoved;
 	}
 
@@ -637,7 +634,7 @@ export class ChartModel implements IDestroyable {
 		return this._serieses;
 	}
 
-	public setAndSaveCurrentPosition(x: Coordinate, y: Coordinate, event: MouseEventHandlerEventBase | null, pane: Pane): void {
+	public setAndSaveCurrentPosition(x: Coordinate, y: Coordinate, event: TouchMouseEventData | null, pane: Pane): void {
 		this._crosshair.saveOriginCoord(x, y);
 		let price = NaN;
 		let index = this._timeScale.coordinateToIndex(x);
